+++ common-service/src/main/java/com/ossovita/commonservice/dto/ReservationDto.java
@@ -6,6 +6,7 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -19,7 +20,7 @@ public class ReservationDto {
 
     public ReservationStatus reservationStatus;
 
-    public double reservationPrice;
+    public BigDecimal reservationPrice;
 
     private LocalDateTime reservationStartTime;
