+++ infrastructure/kafka/src/main/java/com/ossovita/kafka/model/ReservationPaymentResponse.java
@@ -7,6 +7,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
@@ -19,7 +21,7 @@ public class ReservationPaymentResponse {
 
     private ReservationPaymentType reservationPaymentType;
 
-    private double reservationPaymentAmount;
+    private BigDecimal reservationPaymentAmount;
 
     private long reservationFk;
